+++ core/org.eclipse.cdt.core/parser/org/eclipse/cdt/internal/core/index/composite/cpp/CompositeCPPField.java
@@ -41,7 +41,7 @@ class CompositeCPPField extends CompositeCPPVariable implements ICPPField {
 	}
 
 	@Override
-	public byte getFieldPosition() {
+	public int getFieldPosition() {
 		return ((ICPPField)rbinding).getFieldPosition();
 	}
 }
