+++ core/org.eclipse.cdt.core/parser/org/eclipse/cdt/internal/core/index/composite/cpp/CompositeCPPFieldTemplate.java
@@ -41,7 +41,7 @@ public class CompositeCPPFieldTemplate extends CompositeCPPVariableTemplate impl
 	}
 
 	@Override
-	public byte getFieldPosition() {
+	public int getFieldPosition() {
 		return ((ICPPField) rbinding).getFieldPosition();
 	}
 }
