+++ core/org.eclipse.cdt.core/parser/org/eclipse/cdt/internal/core/pdom/dom/cpp/PDOMCPPFieldInstance.java
@@ -8,6 +8,7 @@
  *
  * Contributors:
  *     Lukas Wegmann (IFS) - Initial API and implementation
+ *     Sergey Prigogin (Google)
  *******************************************************************************/
 package org.eclipse.cdt.internal.core.pdom.dom.cpp;
 
@@ -53,7 +54,7 @@ public class PDOMCPPFieldInstance extends PDOMCPPVariableInstance implements ICP
 	}
 
 	@Override
-	public byte getFieldPosition() {
-		return ((ICPPField)getSpecializedBinding()).getFieldPosition();
+	public int getFieldPosition() {
+		return ((ICPPField) getSpecializedBinding()).getFieldPosition();
 	}
 }
